anolis: x86: fix compiler warning in zx_apei_mce_report_pcie_error
ANBZ: #9799

Warning while compiling in x86 platform:
arch/x86/kernel/cpu/mce/apei.c: In function ‘zx_apei_mce_report_pcie_error’:
arch/x86/kernel/cpu/mce/apei.c:131:40: warning: left shift count >= width of type [-Wshift-count-overflow]
  131 |  m.misc |= pcie_err->device_id.segment << 32;
      |                                        ^~
So, fix it.

Fixes: 3fe177a5aef6 ("anolis: x86/mce/zhaoxin: Enable mcelog to decode PCIE, ZDI/ZPI, and DRAM errors")

+++ arch/x86/kernel/cpu/mce/apei.c
@@ -128,7 +128,7 @@ void zx_apei_mce_report_pcie_error(int severity, struct cper_sec_pcie *pcie_err)
 	mce_setup(&m);
 	m.addr = 0;
 	m.misc = 0;
-	m.misc |= pcie_err->device_id.segment << 32;
+	m.misc |= (u64) pcie_err->device_id.segment << 32;
 	m.misc |= pcie_err->device_id.bus << 24;
 	m.misc |= pcie_err->device_id.device << 19;
 	m.misc |= pcie_err->device_id.function << 16;
